+++ packages/terra-overlay/src/Overlay.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import FocusTrap from 'focus-trap-react';
+import { Portal } from 'react-portal';
 import 'terra-base/lib/baseStyles';
 import styles from './Overlay.module.scss';
 import Container from './OverlayContainer';
@@ -13,6 +14,8 @@ const BackgroundStyles = {
   DARK: 'dark',
   CLEAR: 'clear',
 };
+
+const zIndexes = ['100', '6000', '7000', '8000', '9000'];
 
 const propTypes = {
   /**
@@ -39,6 +42,10 @@ const propTypes = {
   * Indicates if the overlay is relative to the triggering container.
   */
   isRelativeToContainer: PropTypes.bool,
+  /**
+  * Z-Index layer to apply to the ModalContent and ModalOverlay. Valid values are '100', '6000', '7000', '8000', or '9000'.
+  */
+  zIndex: PropTypes.oneOf(zIndexes),
 };
 
 const KEYCODES = {
@@ -52,6 +59,7 @@ const defaultProps = {
   isScrollable: false,
   isRelativeToContainer: false,
   onRequestClose: undefined,
+  zIndex: '100',
 };
 
 class Overlay extends React.Component {
@@ -71,6 +79,7 @@ class Overlay extends React.Component {
 
   componentWillUnmount() {
     document.removeEventListener('keydown', this.shouldHandleESCKeydown);
+    this.resetBeforeOverlay();
   }
 
   setContainer(node) {
@@ -134,7 +143,7 @@ class Overlay extends React.Component {
 
   render() {
     const {
-      children, isOpen, backgroundStyle, isScrollable, isRelativeToContainer, onRequestClose, ...customProps
+      children, isOpen, backgroundStyle, isScrollable, isRelativeToContainer, onRequestClose, zIndex, ...customProps
     } = this.props;
     const type = isRelativeToContainer ? 'container' : 'fullscreen';
 
@@ -143,12 +152,18 @@ class Overlay extends React.Component {
       return null;
     }
 
+    let zIndexLayer = '100';
+    if (zIndexes.indexOf(zIndex) >= 0) {
+      zIndexLayer = zIndex;
+    }
+
     const OverlayClassNames = cx([
       'overlay',
       type,
       backgroundStyle,
       { scrollable: isScrollable },
       customProps.className,
+      `layer-${zIndexLayer}`,
     ]);
 
     /*
@@ -169,15 +184,24 @@ class Overlay extends React.Component {
       return overlayComponent;
     }
 
+    const backgroundScrollContent = (
+      <div className={cx('background-scroll-content')}>
+        <div className={cx('inner')} />
+      </div>
+    );
+
     return (
-      <FocusTrap>
-        {overlayComponent}
-      </FocusTrap>
+      <Portal>
+        {backgroundScrollContent}
+        <FocusTrap>
+          {overlayComponent}
+        </FocusTrap>
+      </Portal>
     );
   }
 }
 
-const Opts = { BackgroundStyles };
+const Opts = { BackgroundStyles, zIndexes };
 
 Overlay.propTypes = propTypes;
 Overlay.defaultProps = defaultProps;