+++ packages/terra-show-hide/CHANGELOG.md
@@ -2,7 +2,17 @@ ChangeLog
 =========
 
 Unreleased
------------------
+----------
+
+1.2.0 - (October 30, 2018)
+------------------
+### Changed
+* Update mixin import to be more explicit
+
+1.1.0 - (October 24, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 
 1.0.0 - (October 19, 2018)
 ------------------